Rename files to match renamed class
