+++ compiler/AST/CMakeLists.txt
@@ -56,8 +56,8 @@ set(SRCS
     primitive.cpp
     stmt.cpp
     symbol.cpp
+    TemporaryConversionThunk.cpp
     TransformLogicalShortCircuit.cpp
-    Thunk.cpp
     TryStmt.cpp
     type.cpp
     UseStmt.cpp
