+++ compiler/AST/baseAST.cpp
@@ -43,7 +43,7 @@
 #include "stmt.h"
 #include "stringutil.h"
 #include "symbol.h"
-#include "Thunk.h"
+#include "TemporaryConversionThunk.h"
 #include "TryStmt.h"
 #include "type.h"
 #include "WhileStmt.h"
